requirements.txt
fastapi
uvicorn
aiohttp
beautifulsoup4
playwright
google-genai
pydantic
python-dotenv